+++ src/TestTrackingDiagrams.Extensions.MassTransit/MassTransitOperation.cs
@@ -1,0 +1,12 @@
+namespace TestTrackingDiagrams.Extensions.MassTransit;
+
+public enum MassTransitOperation
+{
+    Send,
+    Publish,
+    Consume,
+    SendFault,
+    PublishFault,
+    ConsumeFault,
+    Other
+}