+++ src/TestTrackingDiagrams.Extensions.MassTransit/MassTransitOperationInfo.cs
@@ -1,0 +1,9 @@
+namespace TestTrackingDiagrams.Extensions.MassTransit;
+
+public record MassTransitOperationInfo(
+    MassTransitOperation Operation,
+    string? MessageType,
+    Uri? DestinationAddress = null,
+    Uri? SourceAddress = null,
+    Guid? MessageId = null,
+    Guid? ConversationId = null);