+++ src/TestTrackingDiagrams.Extensions.MassTransit/MassTransitTrackingOptions.cs
@@ -1,0 +1,15 @@
+namespace TestTrackingDiagrams.Extensions.MassTransit;
+
+public record MassTransitTrackingOptions
+{
+    public string ServiceName { get; set; } = "MassTransit";
+    public string CallingServiceName { get; set; } = "Caller";
+    public MassTransitTrackingVerbosity Verbosity { get; set; } = MassTransitTrackingVerbosity.Detailed;
+    public Func<(string Name, string Id)>? CurrentTestInfoFetcher { get; set; }
+    public Func<string?>? CurrentStepTypeFetcher { get; set; }
+    public bool TrackSend { get; set; } = true;
+    public bool TrackPublish { get; set; } = true;
+    public bool TrackConsume { get; set; } = true;
+    public bool LogMessageBody { get; set; } = true;
+    public bool LogFaults { get; set; } = true;
+}