+++ src/TestTrackingDiagrams.Extensions.MassTransit/MassTransitTrackingVerbosity.cs
@@ -1,0 +1,8 @@
+namespace TestTrackingDiagrams.Extensions.MassTransit;
+
+public enum MassTransitTrackingVerbosity
+{
+    Raw,
+    Detailed,
+    Summarised
+}